addeds: module utils and is_uuid checker
