+++ app/handlers/personal/__init__.py
@@ -1,7 +1,7 @@
 from aiogram import Dispatcher
 
 # тут додавати импорт нових регістраторов
-from app.handlers.personal.cmd_settings import register_handler_settings
+from app.handlers.personal.user_settings.cmd_settings import register_handler_settings
 from app.handlers.personal.start import register_handler_start
 from app.handlers.personal.study import register_handler_study
 from app.handlers.personal.quick_self_test import register_handler_quick_selt_test
