+++ app/utils/__init__.py
@@ -1,0 +1,1 @@
+from utils import is_uuid
