+++ app/utils/utils.py
@@ -1,0 +1,10 @@
+from uuid import UUID
+
+
+def is_uuid(str: str) -> bool:
+    """Check if value is uuid"""
+    try:
+        UUID(str, version=4)
+        return True
+    except Exception:
+        return False
